Added PDAF initialisation step. This includes ensemble initialisation and statevector routines. The current implementation requires a LINEAR FREE SURFACE assumption within NEMO. The linear free surface assumption means that only ssh, T, S, u and v are prognostic variables. For applications that REQUIRE A NONLINEAR FREE SURFACE e.g. coastal and tidal applications, the current codebase will have to be extended.
Because of how XIOS is attached to NEMO, some modifications to the PDAF source code are
required. If these modifications are not added, the current code WILL NOT WORK. (There will
be a MPI deadlock situation.) Please contact the PDAF developer to obtain these
modifications.

The various routines and variables for manipulating the statevector are contained in
mod_statevector. Some of these routines (e.g. initialising the ensemble) are just templates.
These templates will have to be developed according to each individual project.

The callback routine prepostep_ens_pdaf is also currently a template. This will need to be
extended for outputting files after the analysis step. (It is planned to include routines
for outputting such files in a later commit.)