Added PDAF ensemble step. Future work *SHOULD NOT* be branched from this commit as an incremental analysis update (IAU) is required, which is not yet implemented. The IAU method will be implemented in the next commit.
The implementation is designed for LETKF. It should be possible to modify for other filters with little
or no difficulty. A basic module for assimilating SSH observations available on the model grid with added
noise (identical twin experiments) has been included. Unfortunately, this module is not particularly
helpful for assimilating observations not available on the model grid (a template module for this purpose
will be added in a later commit).

Horizontal localization is assumed, and local domains are defined as all those points where the surface
Tmask is ocean. Any points where the surface Tmask is land are not used.

An edge case to be aware is where a local PE has no valid local domains (see init_n_domains_pdaf).

The current implementation *IS NOT* valid for the NEMO diurnal timestepping configuration.

Output on ARCHER2 machine is bit-identical for ORCA2_ICE_PISCES test-case i.e. NEMO-PDAF simulations
*without* assimilation are bit-identical to stand-alone NEMO simulations for ORCA2_ICE_PISCES test-case.