Added incremental analysis update (IAU) step. It is now possible to perform ensemble data assimilation experiments with NEMO-PDAF (although diagnostic functionality such as saving the increments to file is not yet implemented, this functionality will be added in a later commit). Future projects can be safely branched from this commit onwards (although see caveats below).
The IAU step follows closely the implementation of NEMOVAR. However, there are some features which still
need to be added/designed. The most important feature which has not yet been added is the weighting
function for the IAU. This should be quite straightforward to add, and so this is left to future users
to implement.

Another important caveat for this step is that it is *ONLY VALID* for a linear, filtered free surface
assumption. It should be possible to extend this implementation to a more general case (e.g. non-linear
surface with time-splitting), and some basic guidelines have been included in the code on what should
be modified.

Output on ARCHER2 machine is bit-identical for ORCA2_ICE_PISCES test-case i.e. NEMO-PDAF simulations
*without* assimilation are bit-identical to stand-alone NEMO simulations for ORCA2_ICE_PISCES test-case.